Wired up libraries and Joke button
All libraries are wired up correctly now to MainActivity so that when
the joke button is pressed, an AsyncTask calls our backend api to get
a random joke (using the JokeLibrary Java library). The result is sent
back to MainActivity through an interface, and MainActivity starts a
new activity (from the JokeActivityLibrary) using the result of the
AsyncTask.